requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
python-docx
fpdf2
pillow
speechrecognition
openpyxl
pandas
ffmpeg-python
pytesseract
reportlab
docx2pdf
PyMuPDF
pyttsx3

# Font desteği için
fonttools  # TTF font dosyalarının okunabilirliğini artırır
